Make all imports relative
I've toyed around with using the deno.json a lot and
unfortunately I can't seem to get it to work. I'm getting issues with my IDE being
able to recognize file imports and then when I tried to push the first release and
utilize it in a different script there was a relative import path error because it
was not prefixed with slash or dot slash so I'm going to deal with this later.